requirements.txt
ansicon==1.89.0
arrow==0.15.5
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
blessed==1.17.4
cachetools==4.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
constantly==15.1.0
coverage==5.0.4
cryptography==2.8
cssselect==1.1.0
Django==3.0.5
django-background-tasks==1.2.5
django-compat==1.0.15
django-filter==2.2.0
django-guardian==2.2.0
django-picklefield==2.1.1
django-q==1.2.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
fake-useragent==0.1.11
google-api-core==1.17.0
google-auth==1.14.3
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
jinxed==1.0.0
jproperties==2.1.0
json-utils==0.2
jsonpickle==1.4.1
lxml==4.5.0
Markdown==3.2.1
mysqlclient==1.4.6
parsel==1.5.2
Protego==0.1.16
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2019.3
queuelib==1.5.0
redis==3.4.1
requests==2.23.0
rsa==4.0
scraperapi-sdk==0.2.2
Scrapy==2.0.1
scrapy-djangoitem==1.1.1
scrapy-fake-useragent==1.2.0
scrapy-splash==0.7.2
Scrapy-UserAgents==0.0.1
scrapyd==1.2.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
zipp==3.1.0
zope.interface==5.0.2
